Relicense to GPLv3+ because of Mutagen dependency; add code to facilitate non-GPL use as a library
Mutagen has always been licensed under the GPL, but I did not check when
I started this project in 2010 (when I was 18, IIRC).

I've added an exception to allow use under the X11 License if the
Mutagen dependency is removed, or if a portion which does not depend
on Mutagen is used.  The exception also applies when Leviathan is
imported and used as a library without calling its `enable_gpl()`
function.

The above arrangement (i.e. the LICENSE.txt from this commit) also
applies to all previous commits to the extent possible, except that:

a. Prior to Mutagen commit 598f0be6b5ed0fe18e8bf0560c54a9c070754d79
   (2017-02-13), Mutagen was licensed under GPLv2 only, and so that
   license replaces GPLv3 or later when Leviathan uses versions of
   Mutagen that were published prior to that commit.

b. The `enable_gpl()` function was added in this commit, so versions
   of Leviathan prior to this commit must always be used under the GPL.

My apologies to the Mutagen developers.